[B2BP-178] Review pre-header component (#73)
* Revert "commit branch updates"

This reverts commit c62bb59c3a05395f78277dfeaa8af51f904f9c86.

* Change structure of nextjs-website/src/lib

* Add PreHeader component

* Add preHeader fetch

* Add getPreHeaderProps

* Apply PreHeader to layout

* Apply lint changes

* Small fixes

---------
