--- apps/nextjs-website/src/components/reusable/io-ts-declarations.ts
@@ -1,51 +1,0 @@
-import * as t from 'io-ts';
-
-export const ThemeSchema = t.keyof({
-  light: null,
-  dark: null,
-});
-
-export const CTAButtonVariant = t.keyof({
-  text: null,
-  outlined: null,
-  contained: null,
-});
-
-export const CTAButtonColor = t.keyof({
-  inherit: null,
-  primary: null,
-  secondary: null,
-  success: null,
-  error: null,
-  info: null,
-  warning: null,
-});
-
-export const CTAButtonSize = t.keyof({
-  small: null,
-  medium: null,
-  large: null,
-});
-
-export const CtaButtonsSchema = t.intersection([
-  t.type({
-    text: t.string,
-    href: t.string,
-    variant: CTAButtonVariant,
-    color: CTAButtonColor,
-  }),
-  t.partial({
-    icon: t.union([t.string, t.null]),
-    size: CTAButtonSize,
-  }),
-]);
-
-export const CtaGroupCodec = t.intersection([
-  t.type({
-    theme: ThemeSchema,
-    reverse: t.boolean,
-  }),
-  t.partial({
-    ctaButtons: t.array(CtaButtonsSchema),
-  }),
-]);
