+++ apps/nextjs-website/src/lib/api.ts
@@ -1,10 +1,9 @@
 /** This file contains all the functions useful to get data from external resources */
 import { pipe } from 'fp-ts/lib/function';
 import * as E from 'fp-ts/lib/Either';
-import { PreHeaderProps } from '@pagopa/pagopa-editorial-components/dist/components/PreHeader';
-import { getNavigation } from './api/navigation/navigationAPI';
-import { Page, makePageListFromNavigation } from './api/navigation/pages';
-import { getPreHeader } from './api/preHeaderAPI';
+import { Page, makePageListFromNavigation } from './pages';
+import { getNavigation } from './fetch/navigation';
+import { PreHeader, getPreHeader } from './fetch/preHeader';
 import { makeAppEnv } from '@/AppEnv';
 
 // create AppEnv given process env
@@ -22,8 +21,12 @@ export const getAllPages = async (): Promise<ReadonlyArray<Page>> => {
   return makePageListFromNavigation(navigation);
 };
 
-// Return data for preHeader
-export const getPreHeaderData = async (): Promise<PreHeaderProps> => {
-  const preHeaderAPIRes = await getPreHeader(appEnv);
-  return preHeaderAPIRes.data.attributes;
+// Return PreHeaderProps
+export const getPreHeaderProps = async (): Promise<
+  PreHeader['data']['attributes']
+> => {
+  const {
+    data: { attributes },
+  } = await getPreHeader(appEnv);
+  return attributes;
 };
