+++ apps/nextjs-website/src/lib/fetch/extractFromResponse.ts
