+++ apps/nextjs-website/src/lib/fetch/types/Theme.ts
@@ -1,0 +1,8 @@
+import * as t from 'io-ts';
+
+export const ThemeCodec = t.keyof({
+  light: null,
+  dark: null,
+});
+
+export type Theme = t.TypeOf<typeof ThemeCodec>;
