+++ apps/nextjs-website/src/lib/pages.ts
@@ -1,6 +1,6 @@
 import { pipe } from 'fp-ts/lib/function';
 import * as RA from 'fp-ts/lib/ReadonlyArray';
-import { Navigation } from './navigationAPI';
+import { Navigation } from './fetch/navigation';
 
 export type Page = {
   readonly slug: ReadonlyArray<string>;
